+++ configs/db/tables/gems__tracks.30.sql
@@ -2,6 +2,7 @@
 CREATE TABLE if not exists gems__tracks (
         gtr_id_track                int unsigned not null auto_increment,
         gtr_track_name              varchar(40) CHARACTER SET 'utf8' COLLATE 'utf8_general_ci' not null unique key,
+        gtr_external_description    varchar(100) CHARACTER SET 'utf8' COLLATE 'utf8_general_ci',
 
         gtr_track_info              varchar(250) CHARACTER SET 'utf8' COLLATE 'utf8_general_ci',
         gtr_code                    varchar(64) CHARACTER SET 'utf8' COLLATE 'utf8_general_ci' null,